Add Chrome support
In Chrome, cross-origin image fetching must be done in a background script, so
we need to significantly refactor the extension to support this.

- Move perceptual hash functions into a separate module

- Create a background script that accepts messages, fetches the corresponding
  thumbnails, and sends a response with the computed hash

- Create a shim for the phash module that sends a message to the background
  script and returns the parsed response

When bundling the extension for Firefox, we include the perceptual hash module
as a content script, whereas for Chrome we add the shim as a content script and
specify the phash module and message listener module as background scripts.